maple_dsds: Add IMS support for China Unicom and China Telecom
These carriers do not support 3g, only cdma and 4g. We don't support cdma and because of that voice calls won't work for them unless VoLTE is enabled.

In stock, only China Mobile is supported for IMS. Let's enable support for lilac, maple_dsds and poplar_dsds.

It was reported that using the ir51_ir92_ims_gte will make IMS work for these carriers.

So, to enable IMS for them, the following modifications are done:
* Map China Unicom simid's (S285.1 – S285.3) to a new custid (928)
* Map China Telecom simid's (S8760.1 – S8760.2) to a new custid (929)
* Add a modem.conf file containing ir51_ir92_ims_gte for custid 928 and 929 and copy them to system
* Enable VoLTE in CarrierConfig overlays for these carriers